+++ src/scratch.h
@@ -23,9 +23,11 @@ typedef struct secp256k1_scratch_space_struct {
 
 typedef struct secp256k1_scratch_space_struct secp256k1_scratch_space;
 
+#ifndef SECP256K1_NO_MALLOC
 static secp256k1_scratch* secp256k1_scratch_create(const secp256k1_callback* error_callback, size_t max_size);
 
 static void secp256k1_scratch_destroy(const secp256k1_callback* error_callback, secp256k1_scratch* scratch);
+#endif
 
 /** Returns an opaque object used to "checkpoint" a scratch space. Used
  *  with `secp256k1_scratch_apply_checkpoint` to undo allocations. */